We need to find a way to unify the encoding of the information in the test and training datasets. This will allow us to use the neural net on the test data set properly.
I suggest maybe we merge the two datasets, perform the encoding, and then split the data back into test and train, then we use it to train the NN and then simultaneouly test.

I am currently training neural nets with this encoding model so I hope we can find a way to just use the trained neural nets directly.